Move SharedClassicSparkSession to classic.SharedSparkSession
